+++ src/components/EditSections/EditUserRoles/EditUserRoles.js
@@ -6,21 +6,20 @@ import PropTypes from 'prop-types';
 import { useStripes } from '@folio/stripes/core';
 import { isEmpty } from 'lodash';
 import { FieldArray } from 'react-final-form-arrays';
-import { OnChange } from 'react-final-form-listeners';
 import { useUserTenantRoles, useAllRolesData } from '../../../hooks';
 import UserRolesModal from './components/UserRolesModal/UserRolesModal';
 import { filtersConfig } from './helpers';
 
-function EditUserRoles({ match, accordionId, form:{ change }, initialValues }) {
+function EditUserRoles({ match, accordionId, form:{ change }, assignedRoleIds }) {
   const [isOpen, setIsOpen] = useState(false);
   const [unassignModalOpen, setUnassignModalOpen] = useState(false);
-  const [assignedRoleIds, setAssignedRoleIds] = useState(initialValues.assignedRoleIds);
   const { okapi } = useStripes();
   const intl = useIntl();
   const userId = match.params.id;
 
   const { userRoles, isLoading } = useUserTenantRoles({ userId, tenantId: okapi.tenant });
-  const { data: allRolesData } = useAllRolesData();
+
+  const { isLoading: isAllRolesDataLoading, allRolesMapStructure } = useAllRolesData();
 
   const changeUserRoles = (roleIds) => {
     change('assignedRoleIds', roleIds);
@@ -32,14 +31,14 @@ function EditUserRoles({ match, accordionId, form:{ change }, initialValues }) {
   };
 
   const listItemsData = useMemo(() => {
-    if (isEmpty(assignedRoleIds)) return [];
+    if (isEmpty(assignedRoleIds) || isAllRolesDataLoading) return [];
 
     return assignedRoleIds.map(roleId => {
-      const foundUserRole = allRolesData?.roles?.find(role => roleId === role.id);
+      const foundUserRole = allRolesMapStructure.get(roleId);
 
       return { name: foundUserRole?.name, id: foundUserRole?.id };
     });
-  }, [assignedRoleIds, allRolesData]);
+  }, [assignedRoleIds, isAllRolesDataLoading, allRolesMapStructure]);
 
   const unassignAllMessage = <FormattedMessage
     id="ui-users.roles.modal.unassignAll.label"
@@ -50,7 +49,7 @@ function EditUserRoles({ match, accordionId, form:{ change }, initialValues }) {
     if (isEmpty(fields.value)) return null;
 
     const roleId = fields.value[index];
-    const role = allRolesData?.roles?.find(r => roleId === r.id);
+    const role = allRolesMapStructure.get(roleId);
 
     if (!role) return null;
     return (
@@ -124,12 +123,6 @@ function EditUserRoles({ match, accordionId, form:{ change }, initialValues }) {
         cancelLabel={<FormattedMessage id="ui-users.no" />}
         confirmLabel={<FormattedMessage id="ui-users.yes" />}
       />
-      <OnChange name="assignedRoleIds">
-        {(userAssignedRoleIds) => {
-          const userRoleIds = isEmpty(userAssignedRoleIds) ? [] : userAssignedRoleIds;
-          setAssignedRoleIds(userRoleIds);
-        }}
-      </OnChange>
     </div>
   );
 }
@@ -138,7 +131,7 @@ EditUserRoles.propTypes = {
   match: PropTypes.shape({ params: { id: PropTypes.string } }),
   accordionId: PropTypes.string,
   form: PropTypes.object.isRequired,
-  initialValues: PropTypes.object.isRequired
+  assignedRoleIds: PropTypes.arrayOf(PropTypes.string)
 };
 
 export default withRouter(EditUserRoles);